fix: minor issues (#172)
* 🔧 chore(deepsource): update max_line_length to 88 in DeepSource configuration
The maximum line length is adjusted to 88 to align with common style guides, promoting better readability and consistency across the codebase.

* 🐛 fix(Dockerfile, traefik.Dockerfile): add --no-cache-dir option to pip install commands to reduce image size
The `--no-cache-dir` option is added to the `pip install` commands to prevent pip from caching the installed packages, which helps to reduce the final image size. This change improves the efficiency of the Docker images by minimizing unnecessary data storage.

* 🔧 chore(pyproject.toml): update ruff linting rules to include additional checks and improve code quality
The linting rules in the `pyproject.toml` file are expanded to include
additional checks such as flake8-async, flake8-debugger, and others.
This enhances code quality by enforcing more comprehensive linting
standards and ensuring adherence to best practices across the codebase.

* 🔧 refactor(streamlit_app.py): simplify URL checks and improve code readability
✨ feat(streamlit_app.py): enhance audio compression and speaker identification logic
The URL checks for YouTube links are now simplified using a tuple, improving readability. The audio compression function includes a comment for clarity on the ffmpeg path. The speaker identification prompt is adjusted for better formatting, and the loop for generating speech now uses `enumerate` for cleaner indexing. These changes enhance code maintainability and readability.